# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Supreme Wallet API Documentation",
        "description": "API documentation for Supreme Wallet application",
        "contact": {
            "name": "API Support",
            "email": "cccanoks@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/wallets": {
            "get": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Display a listing of all wallets",
                "description": "Display a listing of all wallets",
                "operationId": "getWallets",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of all wallets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Wallet"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": "Bearer {}"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Store a newly created wallet",
                "description": "Store a newly created wallet",
                "operationId": "storeWallet",
                "requestBody": {
                    "description": "Wallet creation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "balance"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "balance": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Wallet created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "wallet": {
                                            "$ref": "#/components/schemas/Wallet"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": "Bearer {}"
                    }
                ]
            }
        },
...
        "/api/wallets/search": {
            "get": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Search wallets by name",
                "description": "Search wallets by name",
                "operationId": "searchWallets",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query string",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of matching wallets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Wallet"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": "Bearer {}"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Transaction": {
                "title": "Transaction",
                "description": "Transaction model",
                "required": [
                    "user_id",
                    "amount",
                    "type",
                    "status",
                    "wallet_id",
                    "currency"
                ],
                "properties": {
                    "id": {
                        "description": "Transaction ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "$ref": "#/components/schemas/User/properties/id"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "amount": {
                        "description": "Transaction amount",
                        "type": "number",
                        "format": "float",
                        "example": 100.5
                    },
                    "type": {
                        "description": "Transaction type",
                        "type": "string",
                        "example": "deposit"
                    },
                    "status": {
                        "description": "Transaction status",
                        "type": "string",
                        "example": "completed"
                    },
                    "description": {
                        "description": "Transaction description",
                        "type": "string",
                        "example": "Deposit to wallet"
                    },
                    "wallet_id": {
                        "$ref": "#/components/schemas/Wallet/properties/id"
                    },
                    "wallet": {
                        "$ref": "#/components/schemas/Wallet"
                    },
                    "currency": {
                        "description": "Transaction currency",
                        "type": "string",
                        "example": "USD"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "email_verified_at": {
                        "description": "Email verification timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z",
                        "nullable": true
                    },
                    "profile_photo_url": {
                        "description": "Profile photo URL",
                        "type": "string",
                        "example": "https://example.com/photos/profile.jpg"
                    },
                    "transactions": {
                        "description": "User transactions",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Transaction"
                        }
                    },
                    "wallets": {
                        "description": "User wallets",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Wallet"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "Wallet": {
                "title": "Wallet",
                "description": "Wallet model representing a user's financial account",
                "properties": {
                    "id": {
                        "description": "Unique identifier of the wallet",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "$ref": "#/components/schemas/User/properties/id"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "wallet_type_id": {
                        "$ref": "#/components/schemas/WalletType/properties/id"
                    },
                    "wallet_type": {
                        "$ref": "#/components/schemas/WalletType"
                    },
                    "balance": {
                        "description": "Current balance in the wallet",
                        "type": "number",
                        "format": "float",
                        "example": 1000.5
                    },
                    "currency": {
                        "description": "Currency code of the wallet (e.g., USD, EUR, GBP)",
                        "type": "string",
                        "example": "USD"
                    },
                    "transactions": {
                        "description": "List of transactions associated with this wallet",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Transaction"
                        }
                    },
                    "created_at": {
                        "description": "Timestamp when the wallet was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the wallet was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "WalletType": {
                "title": "WalletType",
                "description": "WalletType model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Savings"
                    },
                    "description": {
                        "type": "string",
                        "example": "Savings wallet type"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "wallets": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Wallet"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}